pyproject.toml
[project]
name = "clam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "h5py>=3.11.0",
    "pandas>=2.2.2",
    "pyyaml>=6.0.2",
    "opencv-python>=4.10.0.84",
    "scipy>=1.14.1",
    "scikit-learn>=1.5.2",
    "tqdm>=4.66.5",
    "tensorboardx>=2.6.2.2",
    "matplotlib>=3.9.2",
    "openslide-python>=1.3.1",
    "openslide-bin>=4.0.0.5",
    "pydantic-settings>=2.5.2",
    "timm>=1.0.11",
    "boto3>=1.35.0",
    "lancedb>=0.26.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
